+++ electron-builder.privacy-bundled.cjs
@@ -1,0 +1,10 @@
+const { build } = require('./package.json')
+
+module.exports = {
+  ...build,
+  artifactName: '${productName}-${version}-privacy-bundled-win-${arch}.${ext}',
+  extraResources: [
+    ...build.extraResources,
+    'privacy-filter-model/**/*'
+  ]
+}